Move EntityMixinLit to child
