+++ components/header/d2l-consistent-evaluation-lcb-user-context.js
@@ -1,22 +1,17 @@
 import 'd2l-users/components/d2l-profile-image.js';
 import { css, html, LitElement } from 'lit-element';
 import { bodyCompactStyles } from '@brightspace-ui/core/components/typography/styles.js';
+import { EntityMixinLit } from 'siren-sdk/src/mixin/entity-mixin-lit.js';
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 import { RtlMixin } from '@brightspace-ui/core/mixins/rtl-mixin.js';
+import { UserEntity } from 'siren-sdk/src/users/UserEntity.js';
 
-export class ConsistentEvaluationLcbUserContext extends RtlMixin(LitElement) {
+export class ConsistentEvaluationLcbUserContext extends EntityMixinLit(RtlMixin(LitElement)) {
 
 	static get properties() {
 		return {
-			displayName: {
-				attribute: 'display-name',
-				type: String
-			},
-			profileImageHref: {
-				attribute: 'profile-image-href',
-				type: String
-			},
-			token: {
+			_displayName: {
+				attribute: false,
 				type: String
 			}
 		};
@@ -44,14 +39,35 @@ export class ConsistentEvaluationLcbUserContext extends RtlMixin(LitElement) {
 		`];
 	}
 
+	constructor() {
+		super();
+
+		this._setEntityType(UserEntity);
+	}
+
+	set _entity(entity) {
+		if (this._entityHasChanged(entity)) {
+			this._onUserEntityChanged(entity);
+			super._entity = entity;
+		}
+	}
+
+	_onUserEntityChanged(userEntity, error) {
+		if (error || userEntity === null) {
+			return;
+		}
+
+		this._displayName = userEntity.getDisplayName();
+	}
+
 	render() {
 		return html`
 			<d2l-profile-image
-				href="${this.profileImageHref}"
-				token="${this.token}"
+				href="${this.href}"
+				.token="${this.token}"
 				small
 			></d2l-profile-image>
-			<span class="d2l-body-compact d2l-consistent-evaluation-lcb-user-name">${ifDefined(this.displayName)}</span>
+			<span class="d2l-body-compact d2l-consistent-evaluation-lcb-user-name">${ifDefined(this._displayName)}</span>
 		`;
 	}
 }
