+++ components/header/d2l-consistent-evaluation-learner-context-bar.js
@@ -1,16 +1,15 @@
 import './d2l-consistent-evaluation-lcb-user-context.js';
 import { css, html, LitElement } from 'lit-element';
-import { EntityMixinLit } from 'siren-sdk/src/mixin/entity-mixin-lit.js';
-import { ifDefined } from 'lit-html/directives/if-defined.js';
 import { RtlMixin } from '@brightspace-ui/core/mixins/rtl-mixin.js';
-import { UserEntity } from 'siren-sdk/src/users/UserEntity.js';
 
-export class ConsistentEvaluationLearnerContextBar extends (EntityMixinLit(RtlMixin(LitElement))) {
+export class ConsistentEvaluationLearnerContextBar extends RtlMixin(LitElement) {
 
 	static get properties() {
 		return {
-			_displayName: {
-				attribute: false,
+			href: {
+				type: String
+			},
+			token: {
 				type: String
 			}
 		};
@@ -51,32 +50,10 @@ export class ConsistentEvaluationLearnerContextBar extends (EntityMixinLit(RtlMi
 		`;
 	}
 
-	constructor() {
-		super();
-
-		this._setEntityType(UserEntity);
-	}
-
-	set _entity(entity) {
-		if (this._entityHasChanged(entity)) {
-			this._onUserEntityChanged(entity);
-			super._entity = entity;
-		}
-	}
-
-	_onUserEntityChanged(userEntity, error) {
-		if (error || userEntity === null) {
-			return;
-		}
-
-		this._displayName = userEntity.getDisplayName();
-	}
-
 	render() {
 		return html`
 			<d2l-consistent-evaluation-lcb-user-context
-				profile-image-href="${this.href}"
-				display-name="${ifDefined(this._displayName)}"
+				href="${this.href}"
 				.token="${this.token}"
 			></d2l-consistent-evaluation-lcb-user-context>
 		`;
